requirements.txt
# Python 3.10+ recommended (Jetson JetPack 5/6 typically ships 3.10)
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0
numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.8.0
PyYAML>=6.0
tqdm>=4.66.0
onnx>=1.14.0
onnxruntime>=1.16.0
pytest>=7.4.0
